Address PR review feedback
- Extract findHostDistPackageDir, findRootRepoCheckoutDir, fileExists into
  a side-effect-free module (src/host-dist.ts) so scripts/lib can import
  without pulling in harness env var stripping from shared.ts
- Add Number.isFinite validation to parseCliNumber in serve-realm.ts
- Reword PRERENDER_MANAGER_URL comment (port 1 is "expected closed", not
  "unreachable")
